requirements.txt
python-telegram-bot
openai-whisper
translate
python-dotenv
elevenlabs
numpy==1.26.4